Added Entity class
The player and other entities are handled far better than before.

+++ src/main/java/nitrodragon/collision/AABB.java
@@ -14,8 +14,8 @@ public class AABB {
 
     public Collision getCollision(AABB box2) {
         Vector2f distance = box2.center.sub(center, new Vector2f());
-        distance.x = (float) Math.abs(distance.x);
-        distance.y = (float) Math.abs(distance.y);
+        distance.x = Math.abs(distance.x);
+        distance.y = Math.abs(distance.y);
 
         distance.sub(half_extent.add(box2.half_extent, new Vector2f()));
